unzip.py clean up
